requirements.txt
dearpygui==2.0.0
lxml==5.4.0
numpy==2.2.5
pillow==11.2.1
python-docx==1.1.2
typing_extensions==4.13.2
